move all wandb-related configs into new ingredient
